requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.0.1
django-environ==0.11.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
multidict==6.0.5
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
yarl==1.9.4
